+++ starter_projects/example-ios/get_ios_dependencies.sh
@@ -3,10 +3,10 @@
 rm -r ios_dependencies/device
 rm -r ios_dependencies/simulator
 
-wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/3.0.0/ios_dependencies.zip
+wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/3.1.0/ios_dependencies.zip
 unzip -u -o ios_dependencies.zip
 rm ios_dependencies.zip
 
-wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/3.0.0/AIRSDK_patch.zip
+wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/3.1.0/AIRSDK_patch.zip
 unzip -u -o AIRSDK_patch.zip
 rm AIRSDK_patch.zip