+++ starter_projects/example-tvos/get_tvos_dependencies.sh
@@ -2,10 +2,10 @@
 
 rm -r tvos_dependencies/device
 
-wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/3.0.0/tvos_dependencies.zip
+wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/3.1.0/tvos_dependencies.zip
 unzip -u -o tvos_dependencies.zip
 rm tvos_dependencies.zip
 
-wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/3.0.0/AIRSDK_patch.zip
+wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/3.1.0/AIRSDK_patch.zip
 unzip -u -o AIRSDK_patch.zip
 rm AIRSDK_patch.zip